refactor: move logger to util folder
