--- src/logger.ts
@@ -1,4 +1,0 @@
-import { default as pino } from 'pino';
-export const logger = pino({
-	prettyPrint: true
-});
