+++ src/util/logger.ts
@@ -1,0 +1,3 @@
+import { default as pino } from 'pino';
+
+export const logger = pino();
